refactor(probability): isolate finite distribution contracts
